+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
  - Add `colorMode` and `inspectOptions` properties in options for `serverline.init(options)`. #11
+ - Add `ignoreErrors` property in options for `serverline.init(options)`. #13
 
 
 ## [1.3.1] - 2019-08-20